Create README.md
This app is help to guide the user with their call logs. This will work with for the minute pulse calls and not for the second pulse calls. I developed this app for my own usage. I use the post paid which is having the CUG(Closed user group) numbers also. I use this app to know the remaining minutes count for the month. Also, It maintains the old months data. So users can cross check the calls details when they require.